+++ CMakeLists.txt
@@ -7,6 +7,12 @@ project(robot_remote_control VERSION "0.0.9")
 set(BUILD_SHARED_LIBS ON)
 set(CMAKE_BUILD_TYPE Debug)
 
+
+######################################################## CHECK FOR ZeroMQ
+find_package(PkgConfig)
+pkg_check_modules(ZMQ libzmq)
+
+
 ######################################################## CHECK FOR UDT
 find_package(PkgConfig)
 pkg_check_modules(UDT "udt")
