+++ examples/RobotControllerMain.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 #include "RobotController.hpp"
-#include "Transports/TransportZmq.hpp"
+#include <Transports/ZeroMQ/TransportZmq.hpp>
 // #include "Transports/TransportWebSocket.hpp"
 #include <unistd.h>
