+++ src/CLI/CameraViewer.cpp
@@ -3,7 +3,7 @@
  
 
 #include "RobotController.hpp"
-#include "Transports/TransportZmq.hpp"
+#include "Transports/ZeroMQ/TransportZmq.hpp"
 #include "ConsoleCommands.hpp"
 
 using robot_remote_control::TransportSharedPtr;
