+++ src/TransportWrappers/Readme.md
@@ -1,0 +1,9 @@
+# TransportWrappers
+
+A TransportWrapper wraps a transport to additionally process the data before actually sending it.
+This can be used for compression, encryption, and similar.
+
+# Gzip
+
+The gzip wrapper adds gzip compression to the data send.
+
