+++ test/CMakeLists.txt
@@ -28,10 +28,26 @@ set (COMMON_LIBS
 add_executable(test_suite ${COMMON_SOURCE} )
 target_link_libraries(test_suite ${COMMON_LIBS})
 target_compile_definitions(test_suite PUBLIC -DTRANSPORT_DEFAULT)
+target_include_directories(test_suite
+    PUBLIC
+        $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}>/src/Transports
+        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>
+        $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>
+)
+
 
 add_executable(test_suite_ipc ${COMMON_SOURCE} )
 target_link_libraries(test_suite_ipc ${COMMON_LIBS})
 target_compile_definitions(test_suite_ipc PUBLIC -DTRANSPORT_IPC)
+target_include_directories(test_suite_ipc
+    PUBLIC
+        $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}>/src/Transports
+        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>
+        $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>
+)
+
+
+
 
 if(ZLIB_FOUND)
     add_executable(test_suite_gzip ${COMMON_SOURCE})
@@ -40,6 +56,13 @@ if(ZLIB_FOUND)
       robot_remote_control-transport_wrapper_gzip
     )
     target_compile_definitions(test_suite_gzip PUBLIC -DTRANSPORT_DEFAULT_GZIP)
+    target_include_directories(test_suite_gzip
+    PUBLIC
+        $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}>/src/Transports
+        $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}>/src/TransportWrappers
+        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>
+        $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>
+    )
 endif()
 
 if(UDT_FOUND)
@@ -49,6 +72,12 @@ if(UDT_FOUND)
       robot_remote_control-transport_udt
     )
     target_compile_definitions(test_suite_udt PUBLIC -DTRANSPORT_UDT)
+    target_include_directories(test_suite_udt
+    PUBLIC
+        $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}>/src/Transports
+        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>
+        $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>
+    )
 endif()
 
 if(websocketpp_FOUND)
